updated to use new design based on discovery response
