+++ src/identity-providers/messages.json
@@ -39,8 +39,7 @@
     "discoveryEndpoint": "Discovery endpoint",
     "importConfig": "Import config from file",
     "viewMetaData": "View metadata",
-    "showMetaData": "Show metadata",
-    "hideMetaData": "Hide metadata",
+    "advanced": "Advanced",
     "metadataOfDiscoveryEndpoint": "Metadata of the discovery endpoint",
     "authorizationUrl": "Authorization URL",
     "passLoginHint": "Pass login_hint",
@@ -73,6 +72,7 @@
     "allowedClockSkew": "Allowed clock skew",
     "forwardParameters": "Forwarded query parameters",
     "generalSettings": "General settings",
+    "oidcSettings": "OpenId Connect settings",
     "advancedSettings": "Advanced settings",
     "storeTokens": "Store tokens",
     "trustEmail": "Trust Email",
